devices: virtio-input: make config test independent of working directory
The example_custom_config_file_events_eq_default_keyboard_events test
would fail if the test binary ran from a different $PWD than the root of
the devices crate, which happens when the ChromeOS ebuild runs the unit
tests. Include the data in the test binary and write it out to a
temporary file like the other config file test to fix it.

BUG=b:361215918
TEST=FEATURES=test emerge-$BOARD chromeos-base/crosvm

Change-Id: I10fb8ec14de0e11b067fcb29610dc2b1a3fd3776
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6056327
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>