x86_64: make PVMFW location more flexible
Rework the way the protected VM firmware region is allocated in the
memory map so that it does not need to fall immediately past the end of
RAM; instead, allow guest memory to be allocated in the normal way, then
punch a hole in the memory map for the desired PVMFW location.

The general approach used to insert the PVMFW region is a repeated loop
that finds a region overlapping the desired PVMFW location, splits/trims
it as needed to avoid overlapping, and repeat; the loop terminates once
there are no more overlapping regions, at which point it is safe to
insert a new GuestMemory region for the PVMFW.

A follow-up commit will use this flexibility to move the PVMFW to a
different physical address range.

BUG=b:375568070

Change-Id: I3e919c89c4c1c9d2e6e9b6db4cf69982699225e6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6081939
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>