x86_64: factor out common bios setup in arch_memory_regions()
The BIOS region is the same regardless of whether there is RAM above 4G,
so split it out of the if/else.

Also sort the memory regions before returning them; this is mostly a
cosmetic issue (makes the debug logs easier to read), but it also keeps
the regions in the order expected by the current unit tests.

No functional change.

Change-Id: Ic2b4e6079c01ba3275c34a7d85589c72632b6c03
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6085253
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>