base: windows: named pipes: ignore ERROR_MORE_DATA.
On Windows, ERROR_MORE_DATA is not really an error.  It means
there's more data on the pipe, and it can be read in a subsequent
call. Despite this, we've been treating as an error at the named
pipe library level, and then working around it in consumers that
need to allow those errors. This CL does a couple things:
* Switches PipeConnection to ignore ERROR_MORE_DATA and return
  a successful read.
* Remove all handling of ERROR_MORE_DATA in other parts of CrosVM.
* Documents that Unix's version of ERROR_MORE_DATA (as implemented
  by us) is destructive and requires special handling.

BUG=b:296138586
TEST=tested downstream

Change-Id: Ie9d6131fe8043b757466cb1d00743a4e88356943
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6086137
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Noah Gold <nkgold@google.com>