x86_64: pass payload address & size to pvmfw as DTB properties
When running protected VMs on x86, we require them to boot with DTB
anyway, in order to let pvmfw and other AVF components pass various
secrets and other AVF-specific parameters to protected VMs the same way
it is done on aarch64. So we can leverage this fact and pass the VM
payload start and size info from crosvm to pvmfw as DTB properties as
well, the same way it is done on aarch64 (implemented in [1][2]),
instead of introducing a x86-specific protocol for that.

[1] https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4064759
[2] https://android-review.googlesource.com/c/platform/packages/modules/Virtualization/+/2322013

BUG=b:354116267
TEST=run microdroid and check /sys/firmware/devicetree or run microdroid
with --dump-device-tree

Change-Id: Icdacd694d7558d495b346b07493e696e42dab0c1
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6098001
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>