x86_64: update comment about pvmfw ABI
Payload address and size info is now passed to pvmfw via DT.
Additionally passing payload entry address in the %rdi register is still
useful for development purposes (for testing with an unfinished or stub
pvmfw implementation which jumps to the payload without verification)
but should not be used by the final pvmfw implementation. So update the
comment accordingly.

BUG=b:354116267
TEST=n/a

Change-Id: Id42b272c8a092fb57265d5560c5bd9737cf7f933
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6098002
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>