devices: virtio: balloon: fix --lock-guest-memory
https://crrev.com/c/5251093 moved the memory freeing to the main crosvm
processes, but that is incompatible with --lock-guest-memory. The fix is
to track whether the memory was created in the "locked" mode and
fallback to the old behavior.

This won't work for configurations that must delegate to the main
process, like Windows and Gunyah, but they don't use
--lock-guest-memory.

BUG=b:315173520

Change-Id: I04e5c9b5fe56cae4332c127457267e693ebfb915
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6154710
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Alexandre Marciano Gimenez <raging@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>