x86_64: bzImage: load entire file, including setup code
This will enable code that needs the entire contents of the file
(including the real-mode setup code) while keeping the normal kernel
payload and entry point at the same address as previously. For example,
in the case of PVMFW, the `kernel-address` and `kernel-size` devicetree
properties will point to the entire kernel as originally stored in the
bzImage file on disk, so AVB signature checking can work correctly.

BUG=b:389669040
BUG=b:372755978

Change-Id: I93a1f6483fa1f834a604c7158b10560d63850d6e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6173189
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>