gdb: Don't fail on empty VcpuDebugStatus::RegValue
As an empty result represents a valid register currently unavailable,
don't return an error when the (null) size of the result doesn't match
the expected size of the register.

This fixes the following error:

  [ERROR crosvm::crosvm::gdb] Register size mismatch in RegValue: 8 != 0

when reading a system register that is not available e.g.

  (gdb) info registers OSDTRTX_EL1
  Could not fetch register "OSDTRTX_EL1"; remote failure reply 'E79'

After this patch:

  (gdb) info registers OSDTRTX_EL1
  OSDTRTX_EL1    <unavailable>

BUG=b:389738737

Change-Id: I2e8be86fb8dc15d699a2cea830728b2b99fb355c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6169613
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Pierre-Clément Tosi <ptosi@google.com>