tools: bindgen: clean up kernel for all ARCHs
Previously, only the host (x86) architecture was cleaned up after
building the kernel headers. This leaves behind the arm64 and riscv
intermediate files, and trying to build the ChromeOS kernel from the
same tree will fail with an error because the working copy is not clean.

Change-Id: Ie7585e48b30689d5078ebddda9f4568b300b2b3d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6177002
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>