aarch64: revert DT kernel-size to only include file size
The kernel-size property in the devicetree was unintentionally modified
by commit 49013a5a6d86 ("fdt: pass AddressRange to create_config_node()"),
which caused the reported size to include regions that exist in memory
(such as the .bss section) but not in the image file, breaking some
software that relied on the reported kernel-size to match the size of
the file on disk exactly. Roll back to the original behavior for now.

BUG=b:389759119

Change-Id: If5c9365337fd50c353d2dd61b71fbf94b8bf115c
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6191764
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>