devices: bat:  Rate-limit powerd DBus requests
The battery device initializes its state by retrieving data from powerd
via DBus. If the connection to powerd was broken for any reason, the
current implementation repeatedly attempts to send the DBus requests,
resulting in D-Bus LimitsExceeded error.

With this CL, crosvm avoids sending a request if it sent a DBus call
last 1 second. Also we add unit tests for this behavior.

This is a follow-up of crrev.com/c/5918906.

BUG=b:392811615, b:394543244
TEST=CQ

Change-Id: Id85dbd93e97253e7f55041e098f9bbedaa6e88d8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6210458
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Yuan Yao <yuanyaogoog@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>