aarch64: hypervisor: replace SVE cap from Hypervisor to VM
The initial code for enabling SVE via the auto flag would check if SVE
is supported on the hypervisor. The result was the KVM ioctl advertising
the SVE capability, but a VM could be built without SVE support, which
would cause the VM ioctl to not advertise the SVE option, resulting in
failure of creation of the VM.

KVM API documentation suggests using the VM ioctl to adapt specifically
to the VM created. This patch replaces the KVM ioctl with the VM ioctl
call.

BUG=b:376260018
TEST=AOSP

Change-Id: I5062ee188131b3863a5b64c68c0348e471747476
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6215250
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Auto-Submit: Elie Kheirallah <khei@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>