devices: virtio-blk: remove flaky wait_timeout in unit test
Rather than waiting for only 300 milliseconds with wait_timeout() in
resize_with_multiple_workers unit test, which could be exceeded if the
system is busy, just use an infinite wait() so the test will wait as
long as needed. If a change ever causes this test to hang, it will still
be noticed (the test runner infrastructure will catch it).

Change-Id: I156bc9236ac28a826e19276ab90ce29ac2fa7540
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6215196
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>