vm_memory: track --file-backed-mapping RAM regions in GuestMemory
`--file-backed-mapping` allows mapping a file in place of parts of guest
RAM, but the bookkeeping isn't quite right. For example, on aarch64, the
crosvm generated device tree does not advertise the file-backed regions
as part of the memory node. This change fixes that issue by tracking the
file-backed regions as part of the `GuestMemory` data structure.

It now becomes important to distinguish RAM and MMIO file-backed
mappings. We could automatically detect if a mapping falls within RAM,
but the definition of the RAM layout is arch and cmdline dependent, so,
to reduce plumping and allow more validation, we now require the mapping
to be declared as either RAM or MMIO by the user on the cmdline.

BUG=b:368143426,b:381972137

Change-Id: Ib3beb8e72f2bf88e8fa7e83533ee53c2075067ba
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6228396
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>