aarch64: fix off-by-one errors in kernel end calculations
The initrd and devicetree blob are placed relative to the end of the
loaded kernel's AddressRange, but this was incorrectly making use of the
`end` field as if it was exclusive (i.e. kernel_end was assumed to be
the first address after the kernel, rather than the address of the last
byte of the kernel). This was coincidentally accounted for by a similar
off-by-one mistake in the ELF kernel loader, but since we are now fixing
that error, these address calculations must be adjusted to ensure the
initrd and FDT don't overlap the end of the kernel.

BUG=b:354116267

Change-Id: I4e9655e7d9e4b76ef0853de03113540179976845
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6235985
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>