plugin: replace kvm_ioapic_state with a zerocopy-able version
The plugin code wants to access various kvm_sys structs as raw bytes for
the purpose of serialization/deserialization, but in the case of struct
kvm_ioapic_state, this is a problem, since it contains a union and
bitfields. We previously worked around this by manually implementing
the zerocopy traits for the bindgen-generated bitfield types, but this
breaks when trying to upgrade to zerocopy 0.8.

Avoid this by manually defining an equivalently sized and aligned
structure without the problematic features. This will get cleaned up
once we drop the separate `kvm` crate and only have the `hypervisor`
crate left.

BUG=b:372549215
TEST=cargo build --features=plugin

Change-Id: I5fb800793663149f4272514416de6620cfbca7e6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6255228
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Zihan Chen <zihanchen@google.com>