base: windows: remove incorrect MaybeUninit usage
The Event code was using MaybeUninit to initialize a HANDLE value, and
it was being used incorrectly (calling as_mut_ptr() on a temporary
that would be dropped by the time it was used). This didn't really
matter in practice, since event_clone is passed by reference as an
output-only parameter to DuplicateHandle(), but it's simpler to just
initialize the HANDLE to INVALID_HANDLE_VALUE (-1) instead.

Fixes a new warning in Rust 1.84:

"error: a dangling pointer will be produced because the temporary
`std::mem::MaybeUninit<libc::c_void>` will be dropped"

Change-Id: I1b56138c4605a21755f2f27e4e772142630157a8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6266468
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>