devices: vhost-vsock: simplify set_vring_{call,kick}
Previously, the vsock device would construct an Interrupt object just to
take a reference to its Event, then discard it. Instead, construct just
the needed Event (which can be converted with no runtime cost from a
File) at the vhost-vsock call sites.

The O_NONBLOCK modification is also only needed for vhost-user devices
that use the kick file descriptor with a cros_async executor, not for
vhost-user vhost-vsock, where the kick fd is just passed along to the
kernel, so that can be applied in the vmm_vhost Backend set_vring_kick()
instance only.

This also allows the only other VhostUserRegularOps::set_vring_call()
and set_vring_kick() callers to be inlined so the functions can be
deleted.

Change-Id: I31bb56636e55c7c276f64122879c976a762f2a26
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6267093
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>