devices: pci: remove unused MsiVary union
There was some code that was initializing the part of the MSI capability
structure that varied based on whether it was a 32-bit or 64-bit cap
etc. but all of the variants actually just end up using the Default impl
which fills the struct with 0s. Replace the union with a simple byte
array of the same size to avoid problems with the upcoming zerocopy 0.8
upgrade, where unions no longer provide .as_bytes().

BUG=b:372549215

Change-Id: Ib71ecb3fa68b2e60aee96a951f220bf18d840c09
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6271886
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>