x86_64: fix potential setup_data overlap of mptable
The MPTable is written at a physical offset of 639 KB, which is inside
the region reserved for the Linux kernel setup_data linked list. This is
mostly a theoretical concern, as setup_data is typically not very large,
but to avoid any possibility of a collision, reduce the end of the
setup_data region so that it no longer overlaps the MPTable.

Also add documentation of the MPTable location to the memory layout
page.

BUG=b:398261493

Change-Id: Ie020bdfbec90f04dcf08d3dbe6261a8d8d0b61a5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6292279
Reviewed-by: Michał Mazurek <mazurekm@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>