devices: virtio-input: optimize config management
The virtio_input_config was being regenerated on the fly for each config
read request. Additionally, the input device snapshot contained a copy
of the whole VirtioInputConfig struct, even though only the select and
subsel fields were mutable; the rest is entirely static and determined
by the way the device is constructed.

Move the select and subsel fields from VirtioInputConfig to Input, and
only regenerate the virtio config when either select or subsel is
changed by the guest driver.

Change-Id: Ice1c777c3ef02ab41a1b4cc9c16a58c5542362f8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6298951
Reviewed-by: Richard Zhang <rizhang@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>