infra: Disable windows builder
The mingw toolchain for rust is broken upstream.
While we wait for a fix, the builder has to be
disabled to unblock an uprev of the rust toolchain.

Note: The windows builder adds very limited coverage
since we are already building and testing mingw64 builds
on linux with wine64. So the cost of disabling is not
too painful.

BUG=b:396467061
TEST=lucicfg validate infra/config/main.star

Change-Id: Ie20c7a0865d67a2ab5314fd3ca0432466a03eb16
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6300293
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Dennis Kempin <denniskempin@google.com>