devices: simplify generate_acpi() interface
Pass the list of SDTs as a mutable reference to a Vec rather than
requiring the device to return the list.

Also ensure that pci_address is available at the VirtioPciDevice layer
so devices don't need to unwrap an Option; it will always be available
by the time generate_acpi() is called.

Change-Id: I85f60cecaa433cb715b5d9b25cb48dc74741444e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6302868
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>