devices: pci: bounds check MSI-X table and PBA accesses
A guest could trigger an out-of-bounds array access by trying to read or
write a MSI-X table entry or pending bit array larger than the supported
number of entries, causing the VMM to panic and exit.

This is a minimal fix using manual bounds checks; ideally, these
functions (and the rest of the MSI-X support code) should be rewritten
using `get()`/`get_mut()` instead of array indexing, but that would be a
larger and more invasive change.

Change-Id: Ide8d113112585757f7313f11acfa759f0167bacd
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6306697
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>