kvm: support KVM_CAP_MSI_DEVID
This is a prereq for KVM_DEV_TYPE_ARM_VGIC_ITS, which will enable MSIs
for aarch64 VMs.

At least for PCI devices, the "device ID" is always the PCI address.
Most of the changes involved plumbing the PCI address around. This is
only needed for aarch64 so far, but I don't want to pollute the code
with too much conditional compilation, so it is done unconditionally and
then ignored inside the hypervisor implementation if not needed.

BUG=b:216355116

Change-Id: I6f5e131c8b7b2fd2575147f1de3a2a87aec8e910
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6325176
Auto-Submit: Frederick Mayle <fmayle@google.com>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>