device: virtio: vsock: pass negotiated queue size to vhost driver
We should be telling the vhost driver the actual size of the queue, not
the max configurable size that crosvm advertised. In practice, the Linux
drivers essentially always negotiate to use the max size, so it is
unlikely someone has encountered this in the wild.

Renamed the constants for clarity.

BUG=b:405127241

Change-Id: I27a8ffd7e70e91a5bacbe143f05131d84cbad91f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6431594
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>