virtcpufreq_v2: Fix up ipc_ratio when host-topology is used
Ipc_ratio is an independent tunable when configuring for a vCPU's
performance when it differs from the host. Since host-cpu-topology
mirrors the host performance with the guest, use the host's capacity to
reflect the the vCPU's performance instead. Also shortcut any need to
create throttle workers that would slow down the vCPU.

Bug: 409618152
Test: ./tools/presubmit
Change-Id: I2093315f03b3b894ed70793f27437820d42d3eb4
Signed-off-by: David Dai <davidai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6442913
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>