e2e_tests: remove lspci_win backcompat file
The end-to-end tests had separate golden output files for lspci that
would be used on Windows and non-Windows hosts, but the only difference
was that the _win version used CR+LF line endings; the textual content
was identical.

Since the backcompat_test_simple_lspci() test code strips '\r' (CR)
characters from the guest output and then directly compares the guest
output with the golden output, these cannot match when the golden output
includes '\r' characters, so this test could not pass on Windows (it is
not currently executed there).

Remove the _win version of the golden output file and use the same
expected lspci output (without '\r') on all platforms. There is no
reason that the guest PCI layout should depend on the host platform.

Change-Id: Iafc9c0b742c19742afa7d69e5c74d2c29e77d9be
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6456666
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>