devices: virtio-balloon: change vq indexes based on features
The virtio specification for the traditional memory balloon device
contains some unclear wording related to virtqueue numbering. The
interpretation crosvm has been using does not match the interpretation
used by the Linux driver, nor the QEMU balloon device, so change crosvm
to match reality. In particular, this affects the page reporting feature
(--balllon-page-reporting), which did not work with the upstream Linux
virtio-balloon driver previously.

The specification assigns fixed numbers to each virtqueue (for example,
vq 2 is statsq, 3 is free_page_vq, and 4 is reporting_vq); however, the
spec also says that these feature-controlled virtqueues only "exist"
when the corresponding feature is "set". The different interpretations
are due to these words, whose meaning is not defined in the spec. crosvm
assumed that this meant that the vq indexes would remain fixed (and the
queues that didn't "exist" would simply be non-functional if the driver
attempted to modify any queue properties or enable the queue), whereas
QEMU and Linux interpreted this to mean that if a device does not
advertise the corresponding feature, that vq is totally removed from the
list and its vq index is reused for the next feature-dependent vq, and
so on for following vqs.

This change modifies crosvm to work like QEMU, so now the advertised
features (note: not the features acknowledged by the driver) control
whether a given queue "exists" or if it should be totally removed and
its index reused for other queues.

TEST: crosvm run --balloon-page-reporting

Change-Id: I76a29744028574ddc5aa31fbb97d0a2f8ece80f2
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6454902
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>