aarch64, riscv64: fdt: remove excess interrupt-map-mask cells
The interrupt-map-mask is supposed to be a single mask that is applied
to an interrupt specifier before it is looked up in the interrupt-map.
It doesn't make sense for it to have one entry per PCI device.

The extra cells were ignored by the Linux interrupt-map parsing code, so
they didn't cause any harm in practice.

BUG=b:410929974

Change-Id: Ia2595d0fc1fb7bd4643b833f5cf289eb32e37a15
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6460473
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>