devices: virtio-video: fix guest region comparison
After the modification in commit 2aca097, the virtio-video device was no
longer functional; it would attempt to compare the backing object of
each memory region with the backing object of the first region in the
list to ensure they were all the same, but the first region's backing
object was a duplicated file handle, so its .as_raw_descriptor() result
was not identical to the original backing object's .as_raw_descriptor().

Fix this by deferring the descriptor clone to the end, after the backing
object comparisons.

Fixes: 2aca097c140d ("vm_memory: return the backing object from `Guest::offset_from_base`")
Change-Id: Ic5b83b98bae75f719ec0ee9f5ad276acd1692c43
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6471315
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>