hypervisor: kvm: remove non-working is_pkvm()
is_pkvm() is supposed to return whether the VM (no matter protected or
non-protected) is running under pKVM, but in fact it returns true for
protected VMs only, causing confusion (see [1]).

is_pkvm() is only used for checking if readonly memslots are supported.
Since this check is not really working correctly, Android is relying on
the out-of-tree crosvm patch [2] instead.

So just remove is_pkvm() and remove this check for the readonly memslot
capability. Note that although pKVM supports readonly slots for
non-protected VMs since android15-6.1 kernel [3] so we could replace
this is_pkvm() check with an explicit check for a protected VM
(KvmCap::ArmProtectedVm), that would be not enough since Android needs
to support older kernels, so it would still require the patch [2].

[1] https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5171308/comment/1c5fdc88_5b98a13f/
[2] https://android-review.googlesource.com/c/platform/external/crosvm/+/2880409
[3] https://android-review.googlesource.com/c/kernel/common/+/2984845

BUG=b:316956218

Change-Id: I7d3f0660c9de68fd7b32deebcc8a1c9707294ea2
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6480181
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>