vmm_vhost: BackendClient: remove unnecessary validation
BackendClient is supposed to be a thin abstraction over the vhost-user
protocol, but it contained a lot of extra checks on the messages being
sent by the caller. All of this checking should be done in the back-end
implementation, not the front-end calling the back-end, and removing the
checks simplifies the code a lot.

Change-Id: I581fe91e368703b8ca4220e7480efad5ea55c583
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6509479
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>