hypervisor: gunyah: Add the size-max property
Add the size-max property in the vm-config node to let Gunyah
know how big the VM's IPA space could be. Setting it to a large
value of 30GB. If there are any allocations to be done by
Gunyah, it can use addresses after this.

Bug: 417242494
Change-Id: Ieb6779684a4514399f9ddbec2163a9c4eb9dc06c
Signed-off-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6618133
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>