vmm_vhost: unix: clean up test_create_failure
The `connect()` helper function made it unclear which of the two
operations was expected to fail; inline the operations at the call sites
so we can call is_err() or unwrap() on them individually.

This also removes a test-only usage of Error::SocketConnect that did not
match the way that error was used outside of the test code.

Change-Id: I4bb56fa4d57ad364cd8ea3cd7895a4e9074e263f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6538069
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>