devices: virtio-scsi: combine Command new() and execute()
The previous design parsed the SCSI CDB into a Command enum, then
immediately called execute() on it. Combine these two steps into a
single execute_cdb() function to avoid the need to construct the
intermediate Command value, allowing the specific CDB to be a reference
instead of copied, and avoid switching on the opcode twice (once on
the original cdb[0] value in Command::new(), and once on the Rust enum
in Command::execute()).

No functional change.

Change-Id: I0ba735bd6c5022a514637721bbd200e8514e61dd
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6553408
Reviewed-by: Joe Hattori <hattorij@google.com>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>