devices: virtio-scsi: clarify max_lba naming confusion
The LogicalUnit had a field named `max_lba`, which actually held the
size of the device in blocks (i.e. the maximum LBA plus one).

Rename it to `last_lba` to make it crystal clear and match the naming
from the SCSI specification; this simplifies the math in various places,
and prevents the error case of a zero-block LUN (which can't be
represented by SCSI READ CAPACITY) by denying such a block device with
an error at startup.

Change-Id: I652bb1a7b1bffb849a4a2d1f469af8c0250fd6cf
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6551982
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Joe Hattori <hattorij@google.com>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>