x86_64: check mptable fits into its region
Previously, we just checked if the end of the MPTable was beyond the end
of memory, but did not verify that it fits within its reserved region.

Convert the MPTABLE_START constant into a range, and add the
corresponding range check so we get a sensible error if the MPTable is
too big, rather than failing one of the later GuestMemory writes with a
confusing error message.

BUG=b:418985706
TEST=crosvm run -c 100

Change-Id: Ibecb0857c3bd6d5fa56999efd1cb31a5699b9157
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6575164
Reviewed-by: Cody Schuffelen <schuffelen@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Elie Kheirallah <khei@google.com>