x86_64: expand MPTable space to allow more vCPUs
This allows space for more per-CPU structures, extending the valid
number of CPUs to the currently-accepted limit of 254.

The MP Table setup code is reordered slightly: we write the MP
Floating Pointer structure first, at a fixed location to ensure it is
within the last kilobyte of low memory, and then the remaining tables
are written to the preceding region.

BUG=b:418985706
TEST=crosvm run -c 254

Change-Id: I45ccd94df0ab2cc9ab2ae64a0f244a7f62879bdd
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6575165
Reviewed-by: Elie Kheirallah <khei@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>