x86_64: fdt: extend the size of flattened dtb
virtual firmware implementations like pvmfw would need to patch and
extend the dtb provided by crosvm. For x86, crosvm passes the dtb in
SETUP_DATA and the size is exactly the same size of the passed dtb.
So pvmfw would not be able to patch it in-place. Extend and zero fill
the dtb area in memory to 16kb so that pvmfw will be able to patch and
and extend the dtb in-place.

This is similar to the aarch64 implementation except for the reserved
size for dtb. aarch64 reserves much more, but x86 do not have too much
of space in SETUP_DATA and do not need that much space as well.

This patch moves the SETUP_DTB as the last entry in SETUP_DATA and
reserves 16kb or rest of the available memory in SETUP_DATA, which ever
is smaller.

BUG=b:423544528
TEST=boot with pvmfw and verify successful boot.

Change-Id: I3b2d2960eb28aa05da917d988eb33bbaa89b0acc
Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6622051
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Dmytro Maluka <dmaluka@chromium.org>