arch: fdt: put pvmfw region in /memory node
This allows the guest OS to reuse the pvmfw region (minus the part
reserved for DICE). This is desirable because we don't yet have a way to
reclaim that memory for use outside the VM and so it is wasted.

BUG=b:431065184

Change-Id: Ie307a92d13e53dff9c518fde8a9e13e8982cc40f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6886686
Auto-Submit: Frederick Mayle <fmayle@google.com>
Reviewed-by: Pierre-Clément Tosi <ptosi@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>